refactor: split RoundDetailPage into hooks + sections
